+++ app/src/main/java/com/gdsc/laundropool/ui/pastPools/PastPoolsViewModel.kt
@@ -1,4 +1,4 @@
-package com.gdsc.laundropool.ui.PastPools
+package com.gdsc.laundropool.ui.pastPools
 
 import androidx.lifecycle.ViewModel
