+++ app/src/main/res/layout/about_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.About.AboutFragment">
+    tools:context=".ui.about.AboutFragment">
 
     <TextView
         android:layout_width="match_parent"
