+++ app/src/main/res/layout/help_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.Help.HelpFragment">
+    tools:context=".ui.help.HelpFragment">
 
     <TextView
         android:layout_width="match_parent"
