+++ app/src/main/res/layout/my_open_pool_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MyOpenPool.MyOpenPoolFragment">
+    tools:context=".ui.myOpenPool.MyOpenPoolFragment">
 
     <TextView
         android:layout_width="match_parent"
