+++ app/src/main/res/layout/past_pools_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.PastPools.PastPoolsFragment">
+    tools:context=".ui.pastPools.PastPoolsFragment">
 
     <TextView
         android:layout_width="match_parent"
